refactor(query-ast): colocate the environments query with its repository
Move the tracing/scores query builders and their compile step out of the
generic query-ast library and into repositories/environments.ts, next to
getEnvironmentsForProject, as file-local functions. Keeps the pure
input->{sql,params} seam (still unit-testable, golden test unchanged) while
removing the cross-file jump and the split write-mode branching. The
query-ast/kysely library keeps only generic internals (nodes, compiler,
plugins, dialect, tenancy, catalog); per-call-site queries live with their
call site now that the Kysely builder is expected to spread into repositories.
